NEW Adds search term for Transaction Table
Adds functionality to search a transaction based on a string search
term, which is searched inside the `Description` of the existing
transactions.

Moves the ReactTable into a more controlled component so we can
properly reset the page when fetching for new transactions with a
search term.

Corrects some tests that were missing `act`.